requirements.txt
# ViTscale Dependencies
# Core ML/DL frameworks
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Vision Transformer library
timm>=0.9.0

# FLOPs calculation
fvcore==0.1.5.post20221221

# Data analysis and visualization
pandas==2.3.0
matplotlib==3.9.4
matplotlib-inline==0.1.7
seaborn==0.13.2
numpy>=1.21.0

# Image processing
Pillow>=9.0.0

# Configuration
PyYAML>=6.0

# Optional but recommended
tqdm>=4.64.0
timm==1.0.15
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
